Create player, computer, and playgame fxn
Created a random number generator to generate a value for the computer.
I used this to pick 1, 2, or 3 and relate it to rock, paper, or
scissors.
I created a player prompt and function to hold its value.
Finally I used these together to create a function that plays one round.
Started working on the final game which involves a for loop.